Create Tor profile's HostContentSettingsMap with setting its is_off_the_record_ to true.
HostContentSettingsMap::is_off_the_record_ is true for the Tor incognito
profile we used while creating Tor windows since it's an incognito profile.
In this commit, We patch the is_tor_profile check when creating the map so the
regular dummy profile we created for Tor will also acts the same as incognito
profiles.
It also solves the issue which content settings changes in the regular profile
are not reflected in Tor profile immediately because
ContentSettingsPref::ReadContentSettingsFromPref will do canonicalization when
|off_the_record| is false which will copy all the prefs from the parent profile
into the OverlayUserPrefStore's ephemeral_user_pref_store_ when creating
ContentSettingsPref, and cause the copied values in the ephemeral_user_pref_store_
are being used later on instead of using the values in persistent_user_pref_store_
which has the latest changes we made in the parent profile.