Move parent_profile_ from BraveProfileImpl to ProfileImpl.
It is possible for service factories to ask for the parent profile before
BraveProfileImpl constructor is run (while we are still running base class's
constructor. For example, ExtensionPrefValueMapFactory was called before
BraveProfileImpl constructor is run so we will access unintialized
parent_profile_ value if we put it in the subclass.